+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Upgraded to Rails 7.2
 - Upgraded Ruby to 3.3.6
 - Added inline student creating & editing in Group view
+- Added database procedure which updates enrollments according to earliest grades
 
 ## 0.34.0 - Group Reports
 - Added Count of Graded/absent/total students and total lesson average mark in lesson view